implemented a client with methods AddCard, Charge
Implemented an API client that allows you to add a card
as well as charge someone for payments.

My usecase here is for lunch time services and might integrate
with Apple Pay since my school's lunchtime queues are one
hell of a mess. With the sample app, I'll be able to have
students reimburse their friends at the front of the queue for
any purchases.

See the examples in example_test.go